refactor: break out http into tests
