+++ tests/end_to_end_cases/mod.rs
@@ -1,4 +1,5 @@
 pub mod flight_api;
+pub mod http;
 pub mod management_api;
 pub mod management_cli;
 pub mod read_api;
